# swagger-spec.json
{"openapi":"3.0.0","paths":{"/users":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"summary":"Create new user","tags":["Users"]}},"/users/me":{"get":{"operationId":"UserController_getProfile","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get current user profile","tags":["Users"]},"put":{"operationId":"UserController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update current user profile","tags":["Users"]},"delete":{"operationId":"UserController_remove","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Delete current user","tags":["Users"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"responses":{"201":{"description":""}},"summary":"User login","tags":["Auth"]}},"/projects":{"post":{"operationId":"ProjectController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Create new project","tags":["Projects"]},"get":{"operationId":"ProjectController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get all user projects","tags":["Projects"]}},"/projects/{id}":{"get":{"operationId":"ProjectController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get project by ID","tags":["Projects"]},"put":{"operationId":"ProjectController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update project","tags":["Projects"]},"delete":{"operationId":"ProjectController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Delete project","tags":["Projects"]}},"/projects/{id}/members":{"get":{"operationId":"ProjectController_getMembers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get all project members","tags":["Projects"]},"post":{"operationId":"ProjectController_addMember","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddProjectMemberDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Add member to project","tags":["Projects"]}},"/projects/{id}/members/{userId}":{"put":{"operationId":"ProjectController_updateMemberRole","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectMemberRoleDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update member role in project","tags":["Projects"]},"delete":{"operationId":"ProjectController_removeMember","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Remove member from project","tags":["Projects"]}},"/projects/{projectId}/tasks":{"post":{"operationId":"TaskController_create","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Create new task","tags":["Tasks"]},"get":{"operationId":"TaskController_findAll","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get all tasks in project","tags":["Tasks"]}},"/projects/{projectId}/tasks/{taskId}/images":{"post":{"operationId":"TaskController_uploadImages","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Upload images for task","tags":["Tasks"]}},"/projects/{projectId}/tasks/{taskId}":{"get":{"operationId":"TaskController_findOne","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get task info","tags":["Tasks"]},"put":{"operationId":"TaskController_update","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update task","tags":["Tasks"]},"delete":{"operationId":"TaskController_remove","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Delete task","tags":["Tasks"]}},"/projects/{projectId}/tasks/{taskId}/tiff":{"get":{"operationId":"TaskController_downloadTiff","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Download task TIFF result","tags":["Tasks"]}}},"info":{"title":"Aerial Monitoring API","description":"API для системы мониторинга с дронов","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com","description":"User email"},"password":{"type":"string","example":"password123","description":"User password"},"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"}},"required":["email","password","firstName","lastName"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"newpassword123"},"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"}}},"CreateProjectDto":{"type":"object","properties":{"name":{"type":"string","example":"My First Project"},"description":{"type":"string","example":"Description of the project"}},"required":["name"]},"UpdateProjectDto":{"type":"object","properties":{"name":{"type":"string","example":"Updated Project Name"},"description":{"type":"string","example":"Updated project description"}}},"AddProjectMemberDto":{"type":"object","properties":{"userId":{"type":"string","example":"user-uuid"},"role":{"type":"string","enum":["OWNER","EDITOR","VIEWER"],"example":"VIEWER","description":"Role in project"}},"required":["userId","role"]},"UpdateProjectMemberRoleDto":{"type":"object","properties":{"role":{"type":"string","enum":["OWNER","EDITOR","VIEWER"],"example":"EDITOR","description":"New role in project"}},"required":["role"]},"CreateTaskDto":{"type":"object","properties":{"name":{"type":"string","example":"Task 1"},"type":{"type":"string","enum":["DRONE_IMAGES","SATELLITE_IMAGES"],"example":"DRONE_IMAGES"}},"required":["name","type"]},"UpdateTaskDto":{"type":"object","properties":{"name":{"type":"string","example":"Updated Task Name"},"type":{"type":"string","enum":["DRONE_IMAGES","SATELLITE_IMAGES"],"example":"DRONE_IMAGES"}}}}}}